Refactor Matrix component and extract sub-components
